+++ IntelliTest.Services/Models/Questions/Closed/ClosedQuestionReviewViewModel.cs
@@ -9,6 +9,6 @@ namespace IntelliTest.Core.Models.Questions.Closed
     public class ClosedQuestionReviewViewModel : ClosedQuestionViewModel
     {
         public int[] CorrectAnswers { get; set; }
-        public decimal Score { get; set; }
+        public float Score { get; set; }
     }
 }
