+++ IntelliTest.Services/Models/Tests/TestResultsViewModel.cs
@@ -16,6 +16,6 @@ namespace IntelliTest.Core.Models.Tests
         public int Grade { get; set; }
         public Mark Mark { get; set; }
         public Guid StudentId { get; set; }
-        public decimal Score { get; set; }
+        public float Score { get; set; }
     }
 }
