+++ IntelliTest.Services/Models/Tests/TestReviewViewModel.cs
@@ -1,9 +1,4 @@
 ﻿using IntelliTest.Core.Models.Questions;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using IntelliTest.Core.Models.Enums;
 using IntelliTest.Core.Models.Questions.Closed;
 
@@ -14,6 +9,6 @@ namespace IntelliTest.Core.Models.Tests
         public List<ClosedQuestionReviewViewModel> ClosedQuestions { get; set; }
         public List<OpenQuestionReviewViewModel> OpenQuestions { get; set; }
         public List<QuestionType> QuestionOrder { get; set; }
-        public decimal Score { get; set; }
+        public float Score { get; set; }
     }
 }
