+++ IntelliTest.Services/Models/Users/StudentViewModel.cs
@@ -12,6 +12,6 @@ namespace IntelliTest.Core.Models.Users
         public Guid Id { get; set; }
         public string Email { get; set; }
         public string ImagePath { get; set; }
-        public List<decimal> TestResults { get; set; }
+        public List<float> TestResults { get; set; }
     }
 }
